Try specifying toolset cxx to bootstrap.sh to get around
+ ./bootstrap.sh
--prefix=/home/conda/feedstock_root/build_artifacts/boost-cpp_1707819196942/_h_env_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_p
--without-libraries=python --with-toolset=gcc
--with-icu=/home/conda/feedstock_root/build_artifacts/boost-cpp_1707819196942/_h_env_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_placehold_p
Building Boost.Build engine with toolset gcc...
Failed to build Boost.Build build engine
Consult 'bootstrap.log' for more details
+ cat bootstrap.log

A C++11 capable compiler is required for building the B2 engine.
Toolset 'gcc' does not appear to support C++11.

** Note, the C++11 capable compiler is _only_ required for building the
B2
+ exit 1
** engine. The B2 build system allows for using any C++ level and any
other
** supported language and resource in your projects.

You can specify the toolset as the argument, i.e.:
    ./build.sh gcc

Toolsets supported by this script are:
    acc, clang, como, gcc, intel-darwin, intel-linux, kcc, kylix,
mipspro,
    pathscale, pgi, qcc, sun, sunpro, tru64cxx, vacpp

For any toolset you can override the path to the compiler with the CXX
environment variable. You can also use additional flags for the compiler
with the CXXFLAGS environment variable.

A special toolset; cxx, is available which is used as a fallback when a
more
specific toolset is not found and the cxx command is detected. The 'cxx'
toolset will use the CXX, CXXFLAGS, and LIBS environment variables, if
present.

Similarly, the cross-cxx toolset is available for cross-compiling by
using the
BUILD_CXX, BUILD_CXXFLAGS, and BUILD_LDFLAGS environment variables to
compile
binaries that will be executed on the build system. This allows CXX etc.
to be
set for cross-compilers to be propagated to subprocesses.